moved gravity form submit out of the api file
